[#175] SearchWritingView UI 구현
