+++ AGAMI/Resources/Assets.xcassets/SorokColor/SButtonDisabled.colorset/Contents.json
@@ -4,10 +4,10 @@
       "color" : {
         "color-space" : "extended-srgb",
         "components" : {
-          "alpha" : "0.750",
-          "blue" : "0xFF",
-          "green" : "0xFF",
-          "red" : "0xFE"
+          "alpha" : "0.450",
+          "blue" : "0x38",
+          "green" : "0x39",
+          "red" : "0x3B"
         }
       },
       "idiom" : "universal"
@@ -22,10 +22,10 @@
       "color" : {
         "color-space" : "extended-srgb",
         "components" : {
-          "alpha" : "0.750",
-          "blue" : "0xFF",
-          "green" : "0xFF",
-          "red" : "0xFE"
+          "alpha" : "0.450",
+          "blue" : "0x38",
+          "green" : "0x39",
+          "red" : "0x3B"
         }
       },
       "idiom" : "universal"
